# Improve performance of getGenericMapResponse